Allow any generic Path as input to register
Previously only String was allowed, which can be annoying for clients
which already have a PathBuf or Path object. The generic type
AsRef<Path> is the standard for APIs which have file path parameters, as
it works across both owned and un-owned paths. Also, AsRef<Path> allows
Strings and OS specific strings to be used directly as well.

The filename parameter itself is only used for calling File::open(),
which itself also takes in an AsRef<Path> as input.